pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "supermetabase"
version = "0.1.0"
description = "MCP server for Metabase integration with Claude"
requires-python = ">=3.10"
license = {file = "LICENSE"}
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
dependencies = [
    "mcp>=1.2.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]
packages = [
    {include = "supermetabase"}
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "ruff>=0.0.280",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["supermetabase"]

[tool.hatch.build.targets.sdist]
include = [
    "supermetabase",
    "tests",
    "pyproject.toml",
    "README.md",
    "LICENSE",
]
